+++ backend/.env
@@ -1,6 +1,6 @@
-SUPABASE_URL=https://your-project.supabase.co
-SUPABASE_JWK_URL=https://your-project.supabase.co/auth/v1/.well-known/jwks.json
-SUPABASE_AUDIENCE=your-project-id
+SUPABASE_URL=https://ggsxpfbssqmcvbcnpjef.supabase.co
+SUPABASE_JWK_URL=https://ggsxpfbssqmcvbcnpjef.supabase.co/auth/v1/.well-known/jwks.json
+SUPABASE_AUDIENCE=ggsxpfbssqmcvbcnpjef
 AWS_REGION=us-east-1
 BEDROCK_MODEL_ID=anthropic.claude-3-5-sonnet-20240620-v1:0
 SYSTEM_PROMPT=你是一個垃圾分類助理，回覆必須嚴格以 JSON 格式。
