+++ backend/tests/conftest.py
@@ -1,7 +1,7 @@
 # tests/conftest.py
 import pytest
 from fastapi.testclient import TestClient
-from app.main import app
+from ..app.main import app
 
 @pytest.fixture
 def client(monkeypatch):
